Making MATAR a submodule
Deleted MATAR header that was bundled with ELEMENTS, and created a MATAR
submodule based on the separate MATAR repository.
Currently, this submodule is hooked up to a branch in my personal fork
of MATAR that contains some very minor compatibility changes needed to
make MATAR a submodule in ELEMENTS.
I hope to get this branch merged into the main branch in the original
repository soon.
Until then, I will keep that branch of my fork up-to-date with
the main branch in the original MATAR repository.