Modified Docs workflow yet again
In the original workflow, I first check out master, build the docs,
install the docs somewhere else in the filesystem, checkout the gh-pages
branch, and copy in the docs, add all, commit, and publish.
The problem was that there were some residual files left over from
master when gh-pages was checked out and these got committed and pushed.
So now I delete the repo after build and install and then clone the repo
again.